refactor into module RuboCop::GraphQL::CompareOrder
